refactored project structure
