+++ discordgo/discordgo.go
@@ -1,4 +1,4 @@
-package discordgolink
+package discordgo
 
 type Discordgolink interface {
