+++ disgo/example/commands.go
@@ -2,18 +2,18 @@ package main
 
 import "github.com/DisgoOrg/disgo/api"
 
-var commands = []api.Command{
+var commands = []*api.CommandCreate{
 	{
-		Name:              "pause",
-		Description:       "pauses the music",
+		Name:        "pause",
+		Description: "pauses the music",
 	},
 	{
-		Name:              "queue",
-		Description:       "shows you all tracks in queue",
+		Name:        "queue",
+		Description: "shows you all tracks in queue",
 	},
 	{
-		Name:              "play",
-		Description:       "plays some music",
+		Name:        "play",
+		Description: "plays some music",
 		Options: []*api.CommandOption{
 			{
 				Type:        api.CommandOptionTypeString,
@@ -25,7 +25,7 @@ var commands = []api.Command{
 				Type:        api.CommandOptionTypeString,
 				Name:        "search-provider",
 				Description: "where to search",
-				Choices: []api.OptionChoice{
+				Choices: []*api.OptionChoice{
 					{Name: "YouTube", Value: "yt"},
 					{Name: "YouTube Music", Value: "ytm"},
 					{Name: "SoundCloud", Value: "sc"},
