+++ disgo/go.mod
@@ -1,0 +1,11 @@
+module github.com/DisgoOrg/disgolink/disgo
+
+go 1.16
+
+replace github.com/DisgoOrg/disgolink => ../
+
+require (
+	github.com/DisgoOrg/disgo v0.3.1
+	github.com/DisgoOrg/disgolink v0.0.0-20210412071636-40769c7951dd
+	github.com/DisgoOrg/log v1.0.3
+)
