+++ disgord/go.mod
@@ -1,0 +1,3 @@
+module github.com/DisgoOrg/disgolink/disgord
+
+go 1.16
